test: relay global object identification
This PR is a continuation of the #3. It introduces tests for global object identification and also a few changes.

I noticed that most people weren't actually *using* the boilerplate but copying-and-pasting files. That wasn't actually my initial idea and for that reason I decided to remove the product module from the source. It will allow anyone to use the boilerplate following the docs to create their own modules (domain based).